importer: missed double conversion

+++ src/importer-requestwriter.c
@@ -676,7 +676,7 @@ static void request_writer(zsock_t *pipe, void *args)
                 statsd_client_count(state->statsd_client, "importer.inserts.count", state->updates_count);
                 prometheus_client_count(IMPORTER_INSERT_COUNT, state->updates_count);
                 statsd_client_timing(state->statsd_client, "importer.inserts.time", state->update_time/1000);
-                prometheus_client_timing(IMPORTER_INSERT_TIME, state->update_time/1000000);
+                prometheus_client_timing(IMPORTER_INSERT_TIME, ((double)state->update_time)/1000000);
                 if (ticks++ % PING_INTERVAL == 0) {
                     // ping mongodb to reestablish connection if it got lost
                     for (int i=0; i<num_databases; i++) {
